refactor(python): unify image_api.rs into image.rs, add memory tests
Merge PyImageApi into image.rs to eliminate cross-module dependency.
Remove Image from root module (only in kornia_rs.image submodule).
Extract shared LUMINANCE_WEIGHTS constant, add 23 zero-copy and
memory isolation tests covering all transforms and augmentations.
